pyproject.toml
[project]
name = "nuclei-graph-transformer"
version = "0.1.0"
authors = [
    { name = "Karolína Rusnačková", email = "525276@mail.muni.cz" },
    { name = "Matěj Pekár", email = "matejpekar@mail.muni.cz" },
]
requires-python = ">=3.12, <3.14"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "torch>=2.6.0",
    "ray>=2.35.0",
    "hydra-core>=1.3.2",
    "mlflow==2.22.0",
    "openslide-python>=1.4.2",
    "rationai-mlkit",
    "rationai-masks",
    "ratiopath>=1.0.1",
    "transformers>=4.57.1",
    "einops>=0.8.1",
    "torchvision>=0.24.0",
]

[tool.uv.sources]
rationai-mlkit = { git = "https://gitlab.ics.muni.cz/rationai/digital-pathology/libraries/mlkit.git" }
rationai-masks = { git = "https://gitlab.ics.muni.cz/rationai/digital-pathology/libraries/masks.git" }
rationai-kube-jobs = { git = "ssh://git@gitlab.ics.muni.cz/rationai/infrastructure/kube-jobs" }

[dependency-groups]
dev = ["mypy", "ruff"]
job = ["rationai-kube-jobs"]
